// frontend/src/app/utilities/types.ts
export enum TypeInviteRole {
  COMMUNITY = 'Comunidad',
  STANDARD = 'Estándar'
}

export enum categoryTypes {
  COMUNIDAD = 'Comunidad',
  DOCUMENTO = 'Documento',
  TIPO_CONTRATO = 'Tipo de contrato',
}

export enum shareOptions {
  CORREO = 'Correo',
  CHAT = 'Chat'
}

export enum privacyOptions {
  PRIVADO = 'Privado',
  PUBLICO = 'Público'
}

export enum notificationTypes {
  COMMUNITY = 'community',
  REQUEST_COMMUNITY = 'request-community',
  PUBLICATION = 'publication',
  USER = 'user',
  FULL_TEXT = 'full_text',
  INVITATION_COMMUNITY = 'invitation',
  CHAT = 'chat'
}

export enum loginProviderTypes {
  GOOGLE_PROVIDER = 'Google',
  REGULAR_PROVIDER = 'Regular'
}

export enum fileTypes {
  DOCUMENT = 'document',
  IMAGE = 'image',
  VIDEO = 'video'
}

export enum extensionFileTypes {
  PDF = 'application/pdf',
  MP4 = 'video/mp4',
  PNG = 'image/png'
}